ci: react-versino
