+++ .github/workflows/ci.yml
@@ -23,14 +23,13 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Install React ${{ matrix.react }}
-        run: npm i react@${{ matrix.react }} react-dom@${{ matrix.react }}
-
-      # `@testing-library/react@13` dropped support for React 16 and 17 so to ensure that we run
-      # fine under React 18 we should upgrade it if we're testing React 18.
-      - name: Install testing-library react for React 18
+      - name: Install React 18 deps
         if: matrix.react == '18'
-        run: npm i @testing-library/react@13
+        run: npm uninstall react-hot-loader && npm i react@${{ matrix.react }} react-dom@${{ matrix.react }} @testing-library/react@14
+
+      - name: Install React <18 deps
+        if: matrix.react == '16' || matrix.react == '17'
+        run: npm i react@${{ matrix.react }} react-dom@${{ matrix.react }}
 
       - name: Run tests
         run: npm test
@@ -44,20 +43,22 @@ jobs:
         node-version: [18.x, 20.x, 21.x]
         react: [16, 17, 18]
 
+    env:
+      REACT_VERSION: ${{ matrix.react }}
+
     steps:
       - uses: actions/checkout@v3
 
       - name: Run visual tests (node ${{ matrix.node-version }})
         run: make ci
 
-      - name: Install React ${{ matrix.react }}
-        run: npm i react@${{ matrix.react }} react-dom@${{ matrix.react }}
-
-      # `@testing-library/react@13` dropped support for React 16 and 17 so to ensure that we run
-      # fine under React 18 we should upgrade it if we're testing React 18.
-      - name: Install testing-library react for React 18
+      - name: Install React 18 deps
         if: matrix.react == '18'
-        run: npm i @testing-library/react@13
+        run: npm uninstall react-hot-loader && npm i react@${{ matrix.react }} react-dom@${{ matrix.react }} @testing-library/react@14
+
+      - name: Install React <18 deps
+        if: matrix.react == '16' || matrix.react == '17'
+        run: npm i react@${{ matrix.react }} react-dom@${{ matrix.react }}
 
       - name: Upload snapshot diffs
         uses: actions/upload-artifact@v3
